Split oauth2 types
