+++ packages/types/src/discord/webhookEvents.ts
@@ -1,9 +1,9 @@
 /** Types for: https://discord.com/developers/docs/events/webhook-events */
 
-import type { OAuth2Scope } from '../discord.js'
 import type { DiscordApplicationIntegrationType } from './applications.js'
 import type { DiscordEntitlement } from './entitlements.js'
 import type { DiscordGuild } from './guilds.js'
+import type { OAuth2Scope } from './oauth2.js'
 import type { DiscordUser } from './users.js'
 
 /** https://discord.com/developers/docs/events/webhook-events#payload-structure */
