+++ packages/types/src/index.ts
@@ -12,6 +12,7 @@ export * from './discord/guildTemplates.js'
 export * from './discord/interactions.js'
 export * from './discord/invites.js'
 export * from './discord/messages.js'
+export * from './discord/oauth2.js'
 export * from './discord/polls.js'
 export * from './discord/reference.js'
 export * from './discord/skus.js'
@@ -21,7 +22,7 @@ export * from './discord/stickers.js'
 export * from './discord/subscriptions.js'
 export * from './discord/users.js'
 export * from './discord/voice.js'
-export * from './discord/webhooks.js'
 export * from './discord/webhookEvents.js'
+export * from './discord/webhooks.js'
 export * from './discordeno.js'
 export * from './shared.js'
